Complete script to produce a simulated data stream
